requirements.txt
backoff
pytz
requests
singer-python
tap-framework